A bit more misc cleanup/organization for chemicals
