+++ src/main/java/mekanism/common/capabilities/GaugeDropperContentsHandler.java
@@ -11,11 +11,11 @@ import mekanism.api.chemical.infuse.IInfusionTank;
 import mekanism.api.chemical.pigment.IPigmentTank;
 import mekanism.api.chemical.slurry.ISlurryTank;
 import mekanism.api.fluid.IExtendedFluidTank;
+import mekanism.common.capabilities.chemical.dynamic.DynamicChemicalHandler.DynamicGasHandler;
+import mekanism.common.capabilities.chemical.dynamic.DynamicChemicalHandler.DynamicInfusionHandler;
+import mekanism.common.capabilities.chemical.dynamic.DynamicChemicalHandler.DynamicPigmentHandler;
+import mekanism.common.capabilities.chemical.dynamic.DynamicChemicalHandler.DynamicSlurryHandler;
 import mekanism.common.capabilities.chemical.dynamic.DynamicChemicalHandler.InteractPredicate;
-import mekanism.common.capabilities.chemical.dynamic.DynamicGasHandler;
-import mekanism.common.capabilities.chemical.dynamic.DynamicInfusionHandler;
-import mekanism.common.capabilities.chemical.dynamic.DynamicPigmentHandler;
-import mekanism.common.capabilities.chemical.dynamic.DynamicSlurryHandler;
 import mekanism.common.capabilities.chemical.item.RateLimitGasHandler.RateLimitGasTank;
 import mekanism.common.capabilities.chemical.item.RateLimitInfusionHandler.RateLimitInfusionTank;
 import mekanism.common.capabilities.chemical.item.RateLimitPigmentHandler.RateLimitPigmentTank;
