+++ src/main/java/mekanism/common/inventory/slot/chemical/SlurryInventorySlot.java
@@ -1,7 +1,6 @@
 package mekanism.common.inventory.slot.chemical;
 
 import com.mojang.datafixers.util.Pair;
-import java.util.Optional;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 import javax.annotation.Nullable;
@@ -16,7 +15,6 @@ import mekanism.api.chemical.slurry.ISlurryTank;
 import mekanism.api.chemical.slurry.Slurry;
 import mekanism.api.chemical.slurry.SlurryStack;
 import mekanism.common.capabilities.Capabilities;
-import mekanism.common.util.MekanismUtils;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
 
@@ -27,13 +25,7 @@ public class SlurryInventorySlot extends ChemicalInventorySlot<Slurry, SlurrySta
 
     @Nullable
     public static ISlurryHandler getCapability(ItemStack stack) {
-        if (!stack.isEmpty()) {
-            Optional<ISlurryHandler> capability = MekanismUtils.toOptional(stack.getCapability(Capabilities.SLURRY_HANDLER_CAPABILITY));
-            if (capability.isPresent()) {
-                return capability.get();
-            }
-        }
-        return null;
+        return getCapability(stack, Capabilities.SLURRY_HANDLER_CAPABILITY);
     }
 
     //TODO: Implement creators as needed
