+++ src/main/java/mekanism/common/tile/base/TileEntityMekanism.java
@@ -52,15 +52,15 @@ import mekanism.common.capabilities.holder.energy.IEnergyContainerHolder;
 import mekanism.common.capabilities.holder.fluid.IFluidTankHolder;
 import mekanism.common.capabilities.holder.heat.IHeatCapacitorHolder;
 import mekanism.common.capabilities.holder.slot.IInventorySlotHolder;
+import mekanism.common.capabilities.resolver.manager.ChemicalHandlerManager.GasHandlerManager;
+import mekanism.common.capabilities.resolver.manager.ChemicalHandlerManager.InfusionHandlerManager;
+import mekanism.common.capabilities.resolver.manager.ChemicalHandlerManager.PigmentHandlerManager;
+import mekanism.common.capabilities.resolver.manager.ChemicalHandlerManager.SlurryHandlerManager;
 import mekanism.common.capabilities.resolver.manager.EnergyHandlerManager;
 import mekanism.common.capabilities.resolver.manager.FluidHandlerManager;
 import mekanism.common.capabilities.resolver.manager.HeatHandlerManager;
 import mekanism.common.capabilities.resolver.manager.ICapabilityHandlerManager;
 import mekanism.common.capabilities.resolver.manager.ItemHandlerManager;
-import mekanism.common.capabilities.resolver.manager.chemical.GasHandlerManager;
-import mekanism.common.capabilities.resolver.manager.chemical.InfusionHandlerManager;
-import mekanism.common.capabilities.resolver.manager.chemical.PigmentHandlerManager;
-import mekanism.common.capabilities.resolver.manager.chemical.SlurryHandlerManager;
 import mekanism.common.config.MekanismConfig;
 import mekanism.common.inventory.container.ITrackableContainer;
 import mekanism.common.inventory.container.MekanismContainer;
