+++ src/main/java/mekanism/common/tile/interfaces/chemical/ISlurryTile.java
@@ -6,10 +6,10 @@ import mcp.MethodsReturnNonnullByDefault;
 import mekanism.api.chemical.slurry.ISlurryTank;
 import mekanism.api.chemical.slurry.Slurry;
 import mekanism.api.chemical.slurry.SlurryStack;
-import mekanism.common.capabilities.chemical.dynamic.DynamicSlurryHandler;
+import mekanism.common.capabilities.chemical.dynamic.DynamicChemicalHandler.DynamicSlurryHandler;
 import mekanism.common.capabilities.chemical.dynamic.ISlurryTracker;
 import mekanism.common.capabilities.holder.chemical.IChemicalTankHolder;
-import mekanism.common.capabilities.resolver.manager.chemical.SlurryHandlerManager;
+import mekanism.common.capabilities.resolver.manager.ChemicalHandlerManager.SlurryHandlerManager;
 import net.minecraft.util.Direction;
 
 @MethodsReturnNonnullByDefault
